requirements.txt
requests==2.31.0
streamlit==1.32.0
google-api-python-client==2.123.0
python-dotenv==1.0.1
reportlab==4.1.0